consumer/requirements.txt
kafka-python>=2.0.2
mysql-connector-python>=8.0.33
python-dotenv>=1.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
